fix: use float 64

+++ src/provider/DirectRenderTasks.ts
@@ -189,7 +189,7 @@ export class DirectRenderTasks {
 
       if (order == null && !this.valueCacheData.has(key)) {
         // build and valueCache
-        const vs = new Float32Array(this.data.length);
+        const vs = new Float64Array(this.data.length);
         let i = 0;
         return this.builder(
           {
